adreno_tz: Fix GPU target frequency calculation for high refresh rates
It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 90 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 120 Hz, is used.

To remedy this, multiply the GPU's busy time by a ratio of the current
refresh rate to 90 Hz. This makes msm-adreno-tz select the right
frequencies needed to avoid stuttering.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mawrick26 <mawrick26@gmail.com>
Tested-by: Avina Kefin <paijo@muslim.com>